requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
ffpyplayer==4.3.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
kiwisolver==1.1.0
matplotlib==3.1.2
moviepy==1.0.0
mtcnn-pytorch==1.0.2
networkx==2.4
numpy==1.16.1
opencv-python==4.1.2.30
pafy==0.5.5
Pillow==6.2.1
proglog==0.1.9
pycryptodome==3.9.4
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.22.0
scikit-image==0.16.2
scipy==1.4.1
six==1.13.0
streamlink==1.3.0
torch==1.3.1
tqdm==4.41.0
urllib3==1.25.7
websocket-client==0.56.0
youtube-dl==2019.11.28